feat: 🎸 Rename model to base model. Refactor assignDefaults
