+++ addons/api/mirage/models/role.js
@@ -1,5 +1,5 @@
 import { belongsTo, hasMany } from 'ember-cli-mirage';
-import Model from './default-model';
+import Model from './base';
 
 export default Model.extend({
   scope: belongsTo(),
